move sound list
