+++ spx-gui/src/components/editor/panels/sound/SoundsPanel.vue
@@ -1,12 +1,39 @@
 <template>
-  <div v-show="props.active" class="sounds-details">Sounds details</div>
-  <div v-show="!props.active" class="sounds-overview">Sounds overview</div>
+  <div v-show="props.active" class="sounds-details">
+    <NInput placeholder="TODO" />
+    <AssetAddBtn :type="AssetType.Sound" />
+    <SoundEditCard
+      v-for="asset in editorCtx.project.sounds"
+      :key="asset.name"
+      :asset="asset"
+      @click="handleSelect(asset)"
+      @remove="handleRemoveSound(asset)"
+    />
+  </div>
+  <div v-show="!props.active" class="sounds-overview"></div>
 </template>
 
 <script setup lang="ts">
+import { AssetType } from '@/apis/asset'
+import { useEditorCtx } from '../../ProjectEditor.vue'
+import type { Sound } from '@/models/sound'
+import AssetAddBtn from '../todo/AssetAddBtn.vue'
+import SoundEditCard from '@/components/editor/sound/SoundEditCard.vue'
+import { NInput } from 'naive-ui'
+
+const editorCtx = useEditorCtx()
+
 const props = defineProps<{
   active: boolean
 }>()
+
+const handleSelect = (asset: Sound) => {
+  editorCtx.select('sound', asset.name)
+}
+
+const handleRemoveSound = (asset: Sound) => {
+  editorCtx.project.removeSound(asset.name)
+}
 </script>
 
 <style scoped lang="scss">
